Move `StoreManagerError` to production
