+++ Sources/Curtz/Store/Store.swift
@@ -1,0 +1,8 @@
+//
+//  Store.swift
+//  CurtzUnitTests
+//
+//  Created by George Nyakundi on 12/05/2023.
+//
+
+import Foundation
